BioBlogsNMalariaCellClassifier/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-crispy-forms==2.1
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.2
protobuf==4.25.3
requests==2.31.0
sqlparse==0.5.0
Pillow==10.3.0
tensorflow==2.16.1
